Extract topic store mutation transaction
